Add a dynamic gradient background
Key presses now add colours dynamically. I still need to add a reset when there are over 6 colours and understand the mapping between the sounds and colours (especially for getting rid of a sound), so I can do 'mock' sound files, all as names. After that, I can add sockets (or even before then if it is easier)